+++ ProjectService/CHANGELOG.md
@@ -1,3 +1,21 @@
+## 0.0.1 (2025-01-28)
+
+
+### Bug Fixes
+
+* moved gitignore to project root ([44c5a42](https://github.com/Travis-Berthrong/Community-Solar/commit/44c5a420a988b9250e9b06759094b110a214692f))
+
+
+### Features
+
+* added testing for ProjectService ([e5d172e](https://github.com/Travis-Berthrong/Community-Solar/commit/e5d172e4f95caa199cc6a7762bc709f68150d972))
+* implemented add investor route ([9d55f43](https://github.com/Travis-Berthrong/Community-Solar/commit/9d55f43793cffe3c5fffa56850383d6ef0e70075))
+* Implemented project controller ([61d0159](https://github.com/Travis-Berthrong/Community-Solar/commit/61d0159a5ff23bb801dfddfa3daab94d77df9b58))
+* implemented project schema and ProjectService ([4363aed](https://github.com/Travis-Berthrong/Community-Solar/commit/4363aedee82d2d3598daee66942a2b8372b3d9ab))
+* setup service structure ([275921d](https://github.com/Travis-Berthrong/Community-Solar/commit/275921dff56c337913d53766acc8f4150e01b579))
+
+
+
 ## 0.0.1 (2025-01-26)